refactor(web): extract messaging-platform catalog into web_routers/messaging (web_server.py god-file slice R4-C2C5)
